% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{gha_debug}
\alias{gha_debug}
\alias{gha_notice}
\alias{gha_warning}
\alias{gha_error}
\title{Add message to the GitHub Actions log}
\usage{
gha_debug(message, frame = caller_env())

gha_notice(
  title,
  message,
  file = NULL,
  line = NULL,
  col = NULL,
  endLine = NULL,
  endCol = NULL,
  frame = caller_env()
)

gha_warning(
  title,
  message,
  file = NULL,
  line = NULL,
  col = NULL,
  endLine = NULL,
  endCol = NULL,
  frame = caller_env()
)

gha_error(
  title,
  message,
  file = NULL,
  line = NULL,
  col = NULL,
  endLine = NULL,
  endCol = NULL,
  frame = caller_env()
)
}
\arguments{
\item{message}{Interpolated glue string giving message to record.
Must be a single line.}

\item{frame}{Frame in which glue will look for variables. Typically only
needed if you are wrapping this in another function.}

\item{title}{Interpolated glue string giving a custom title for the message.}

\item{file, line, col, endLine, endCol}{Optionally, associate a message with
a specific file, and location within that file.}
}
\description{
These messages are highlighted in the log and are displayed on the summary
page.
\itemize{
\item \code{gha_debug()} are only shown when \code{ACTIONS_STEP_DEBUG} is set to \code{true},
which typically happens when you re-run a job and select
"Enable debug logging".
\item \code{gha_notice()} plays a similar role to \code{message()}/\code{cli_inform()}; i.e.
use it for informative notices about things that are working correctly.
\item \code{gha_warning()} plays a similar role to \code{warning()}/\code{cli_warn()}; i.e.
use it to report problems that are being overcome automatically, but
should really be fixed.
\item \code{gha_error()} plays a similar role to \code{stop()}/\code{cli_error()}; i.e. use it
to report problems that require manual intervention.
}
}
