% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{gha_group_start}
\alias{gha_group_start}
\alias{gha_group_end}
\title{Group log lines, optionally adding time taken}
\usage{
gha_group_start(title)

gha_group_end(show_time = TRUE)
}
\arguments{
\item{title}{The title of the group}

\item{show_time}{Show the time taken since the group started?}
}
\description{
GHA allows you to group log lines together, which can be useful for
organizing output. \code{gha_group_start()} starts a groups, and \code{gha_group_end()}
ends it, optionally showing the time taken.
}
