% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{gha_summary}
\alias{gha_summary}
\title{Write markdown to job summary}
\usage{
gha_summary(lines)
}
\arguments{
\item{lines}{A character vector of markdown formatted text.}
}
\description{
This markdown will be displayed at the bottom of the job summary and is
a useful way to deliver rich content to the user.
}
