% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntfy.R
\name{ntfy_send}
\alias{ntfy_send}
\title{Send push notification with ntfy.sh}
\usage{
ntfy_send(id, message, title = "Notification from gh", link = NULL)
}
\arguments{
\item{id}{ID of your notification page.}

\item{message, title}{Message body and title. Interpolated with glue.}

\item{link}{Optional link to open when notification is clicked.}
}
\description{
This is a very lightweight wrapper to for \url{https://ntfy.sh} using the curl
package. We suggest using ntfy.sh because it's free and doesn't require any
setup.
}
\examples{
ntfy_send("hKaZodPEhZVdcMHU", "Hi you! Today is {format(Sys.Date(), '\%A')}.")
browseURL("https://ntfy.sh/hKaZodPEhZVdcMHU")
}
\seealso{
\url{https://github.com/jonocarroll/ntfy} for a more feature-rich
wrapper.
}
