% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha-package.R
\docType{package}
\name{gha-package}
\alias{gha}
\alias{gha-package}
\title{gha: Useful functions for GitHub Actions}
\description{
When running code on GitHub actions, you want to make your code as easy as possible to debug if something goes wrong. This package provides you with tools to achieve that goal.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
